Made the production version run at /obstool/ and added github deployment

+++ ObsTool/Scripts/PROD-build-and-run-integrated.cmd
@@ -1,14 +1,25 @@
 @echo off
 echo ############################################################
-echo #               Deploying and running ObsTool              #
+echo #               Building and running ObsTool               #
 echo ############################################################
 
-cd C:\Users\Olle\source\repos\olleeriksson\ObsTool\ObsTool
-rmdir /s /q "bin"
+set "ASPNETCORE_ENVIRONMENT=Production"
+set "ASPNETCORE_URLS=http://localhost:5000"
+set "ASPNETCORE_PATHBASE=/obstool"
+set "VITE_API_URL=/obstool/api"
+set "VITE_BASE_PATH=/obstool/"
+
+pushd "%~dp0.."
+if exist "bin" rmdir /s /q "bin"
 dotnet clean
 dotnet publish --configuration Release -verbosity detailed
-:: --self-contained :: /p:PublishSingleFile=true
+if errorlevel 1 (
+  popd
+  exit /b %errorlevel%
+)
 
-REM Updated path from netcoreapp3.0 to net10.0 (post-upgrade)
-cd C:\Users\Olle\source\repos\olleeriksson\ObsTool\ObsTool\bin\Release\net10.0\publish
+cd bin\Release\net10.0\publish
+echo.
+echo ObsTool local production URL: http://localhost:5000/obstool/
 .\ObsTool.exe
+popd
